// frontend/src/api-generated/models/thongKeTongQuanDashboardResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */
import type { ThongKeTongQuanDashboardResponseBieuDoPhanBoPhongBan } from './thongKeTongQuanDashboardResponseBieuDoPhanBoPhongBan';
import type { ThongKeTongQuanDashboardResponseBieuDoTyLeTrangThai } from './thongKeTongQuanDashboardResponseBieuDoTyLeTrangThai';

export interface ThongKeTongQuanDashboardResponse {
  idDonVi?: number;
  tongSoLuongThietBi?: number;
  tongGiaTriTaiSanVnd?: number;
  soLuongYeuCauCapPhatChoDuyet?: number;
  soLuongYeuCauBaoHongChoDuyet?: number;
  bieuDoTyLeTrangThai?: ThongKeTongQuanDashboardResponseBieuDoTyLeTrangThai;
  bieuDoPhanBoPhongBan?: ThongKeTongQuanDashboardResponseBieuDoPhanBoPhongBan;
}